factor out cloudkit read operation to view model
