+++ PersonaGenerator/ManagePersonasListView.swift
@@ -10,13 +10,12 @@ import CloudKit
 
 struct ManagePersonasListView: View {
     
-    
-    @State private var personas: [Persona] = []
+    @ObservedObject var viewModel = ViewModel()
     
     var body: some View {
         NavigationView {
             List {
-                ForEach(personas) { persona in
+                ForEach(viewModel.personas) { persona in
                     NavigationLink(destination: EditPersonaView(persona: persona)) {
                         Text(persona.title)
                     }
@@ -29,7 +28,7 @@ struct ManagePersonasListView: View {
                                     // Handle error
                                 } else {
                                     // Remove the deleted record from the `personas` array
-                                    self.personas.removeAll(where: { $0.recordID == recordID })
+                                    viewModel.personas.removeAll(where: { $0.recordID == recordID })
                                 }
                             }
                         }) {
@@ -40,39 +39,14 @@ struct ManagePersonasListView: View {
                 }
             }
             .navigationTitle("Manage Personas")
-            .onAppear {
+            .task {
                 // Fetch data from CloudKit here
-                let privateDatabase = CKContainer.default().privateCloudDatabase
-                let query = CKQuery(recordType: "Persona", predicate: NSPredicate(value: true))
-                privateDatabase.perform(query, inZoneWith: nil) { (records, error) in
-                    if error != nil {
-                        // Handle error
-                    } else {
-                        self.personas = records!.compactMap { record in
-                            guard let title = record["title"] as? String,
-                                  let imageAsset = record["image"] as? CKAsset,
-                                  let name = record["name"] as? String,
-                                  let headline = record["headline"] as? String,
-                                  let bio = record["bio"] as? String,
-                                  let birthdate = record["birthdate"] as? Date,
-                                  let email = record["email"] as? String,
-                                  let phone = record["phone"] as? String,
-                                  let images = record["images"] as? [CKAsset] else {
-                                return nil
-                            }
-                            // Load the image from the CKAsset
-                            let image = UIImage(contentsOfFile: imageAsset.fileURL!.path)
-                            return Persona(recordID: record.recordID, title: title, image: image!, name: name, headline: headline, bio: bio, birthdate: birthdate, email: email, phone: phone, images: images)
-                        }
-                        
-                    }
-                }
+                viewModel.fetchPersonas()
             }
         }
     }
+    
 }
-
-
 
 struct PersonasListView_Previews: PreviewProvider {
     static var previews: some View {
