+++ PersonaGenerator/NetworkSingleton.swift
@@ -8,6 +8,9 @@
 import SwiftUI
 
 class NetworkSingleton {
+    
+    static var shared = NetworkSingleton()
+    
     func fetchImage() -> UIImage {
         var image: UIImage = UIImage(systemName: "person")!
             let url = URL(string: "https://thispersondoesnotexist.com/image")!
