+++ PersonaGenerator/PersonaGenerator/ContentView.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 struct ContentView: View {
+//    @ObservedObject var viewModel: ViewModel
     
     var body: some View {
         TabView {
