+++ PersonaGenerator/PersonaGenerator/PersonaGeneratorApp.swift
@@ -9,6 +9,7 @@ import SwiftUI
 
 @main
 struct PersonaGeneratorApp: App {
+//    @StateObject private var viewModel = ViewModel()
     var body: some Scene {
         WindowGroup {
             ContentView()
